backend/requirements.txt
django
djangorestframework
django-filter
coreapi
pillow
autopep8
mysqlclient
django-admin-interface
django-cors-headers